secmet: split feature file into per-feature files
secmet: separate feature tests into per-class files

fixup: fix smcogs integration test to match new secmet layout
