+++ antismash/common/secmet/feature/__init__.py
@@ -1,0 +1,16 @@
+# License: GNU Affero General Public License v3 or later
+# A copy of GNU AGPL v3 should have been included in this software package in LICENSE.txt.
+
+""" Representations of features within secmet """
+
+from .antismash_domain import AntismashDomain
+from .antismash_feature import AntismashFeature
+from .cds_feature import CDSFeature
+from .cds_motif import CDSMotif
+from .cluster import Cluster
+from .cluster_border import ClusterBorder
+from .domain import Domain
+from .feature import Feature, FeatureLocation, SeqFeature
+from .gene import Gene
+from .pfam_domain import PFAMDomain
+from .prepeptide import Prepeptide
